+++ src/shared/ui/overlay/BottomSheet.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { useRef, useState, type ReactNode } from 'react';
+import type { ReactNode } from 'react';
 import useBodyScrollLock from '@/shared/ui/overlay/hooks/useBodyScrollLock';
+import useBottomSheetDrag from '@/shared/ui/overlay/hooks/useBottomSheetDrag';
 import useEscapeKey from '@/shared/ui/overlay/hooks/useEscapeKey';
 import OverlayBackdrop from '@/shared/ui/overlay/primitives/OverlayBackdrop';
 import OverlayPortal from '@/shared/ui/overlay/primitives/OverlayPortal';
@@ -13,26 +14,11 @@ type BottomSheetProps = {
   onClose: () => void;
   children: ReactNode;
   ariaLabel?: string;
-  /** OverlaySurface에 전달할 추가 className (max-h 등 크기 조정 시 사용) */
   surfaceClassName?: string;
   surfaceStyle?: React.CSSProperties;
   className?: string;
 };
 
-/**
- * 모바일 하단에서 올라오는 바텀시트 공통 컴포넌트입니다.
- *
- * - OverlaySurface `position="bottom" variant="sheet"` 기반
- * - 핸들바를 아래로 100px 이상 드래그하면 닫힙니다.
- * - ESC 키, 백드롭 클릭으로도 닫을 수 있습니다.
- *
- * @example
- * ```tsx
- * <BottomSheet isOpen={isOpen} onClose={onClose} ariaLabel="예약 날짜 선택">
- *   <div className="px-6">콘텐츠</div>
- * </BottomSheet>
- * ```
- */
 export default function BottomSheet({
   isOpen,
   onClose,
@@ -42,36 +28,20 @@ export default function BottomSheet({
   surfaceStyle,
   className,
 }: BottomSheetProps) {
-  const startYRef = useRef<number>(0);
-  const [dragY, setDragY] = useState(0);
-  const handleClose = () => {
-    setDragY(0);
-    onClose();
-  };
-
   useBodyScrollLock({ isLocked: isOpen });
-  useEscapeKey({ isEnabled: isOpen, onEscape: handleClose });
+  useEscapeKey({ isEnabled: isOpen, onEscape: onClose });
 
-  const handleTouchStart = (e: React.TouchEvent) => {
-    startYRef.current = e.touches[0].clientY;
-  };
-
-  const handleTouchMove = (e: React.TouchEvent) => {
-    const diff = e.touches[0].clientY - startYRef.current;
-    if (diff > 0) setDragY(diff);
-  };
-
-  const handleTouchEnd = () => {
-    if (dragY > 100) handleClose();
-    else setDragY(0);
-  };
+  const { dragY, handlers } = useBottomSheetDrag({
+    onClose,
+  });
 
   if (!isOpen) return null;
 
   return (
     <OverlayPortal>
       <div className={cn('fixed inset-0 layer-modal', className)}>
-        <OverlayBackdrop onClick={handleClose} ariaLabel="닫기" />
+        <OverlayBackdrop onClick={onClose} ariaLabel="닫기" />
+
         <OverlaySurface
           position="bottom"
           variant="sheet"
@@ -85,15 +55,11 @@ export default function BottomSheet({
           }}
         >
           <div role="dialog" aria-modal="true" aria-label={ariaLabel} className="flex flex-col">
-            {/* 핸들바 — 드래그 영역 */}
-            <div
-              className="flex shrink-0 justify-center pt-4 pb-3"
-              onTouchStart={handleTouchStart}
-              onTouchMove={handleTouchMove}
-              onTouchEnd={handleTouchEnd}
-            >
+            {/* 핸들바 */}
+            <div className="flex shrink-0 justify-center pt-4 pb-3" {...handlers}>
               <div className="h-1 w-19 rounded-full bg-gray-300" />
             </div>
+
             {/* 콘텐츠 */}
             <div className="flex min-h-0 flex-col overflow-y-auto">{children}</div>
           </div>
